\name{BootPR-package}
\alias{BootPR-package}
\alias{BootPR}
\docType{package}
\title{
 Bootstrap Prediction Intervals and Bias-Corrected Forecasting 
}
\description{
The package provides alternative bias-correction methods for univariate autoregressive model parameters; and generate point forecats and prediction intervals for economic time series.

A future version will include the case of vector AR models.

}
\details{
\tabular{ll}{
Package: \tab BootPR\cr
Type: \tab Package\cr
Version: \tab 0.57\cr
Date: \tab 2009-05-02\cr
License: \tab GPL version 2 or newer\cr
}
}
\author{
Jae H. Kim

Maintainer: Jae H. Kim <Jae.Kim@BusEco.monash.edu.au>

}
\keyword{ts}