\name{IPdata}
\alias{IPdata}
\docType{data}
\title{ US industrial production data}
\description{
  From Extended Nelson-Plosser data set, annua1, 1860-1988
}
\usage{data(IPdata)}


\references{
Andrews, D.W. K., & Chen, H. -Y. (1994). Approximate median-unbiased estimation of autoregressive models. Journal of Business & Economic Statistics, 12, 187-204.
}
\examples{
data(IPdata)
}
\keyword{ts}
